+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 - Core: Fixed link date not displaying in link browser
+- Address book: Added initials field
 
 30-08-2019 6.4.48
 - Core: Install comments and bookmarks by default